Better simulation and removed bug in taxisolver
We were filtering using the opposite condition!

+++ src/entity/Ride.java
@@ -26,7 +26,7 @@ public class Ride {
 		this.points = new ArrayList<>();
 	}
 
-	public double getTravelDistance() { return getStartPosition().getDistance(getEndPosition()); }
+	public int getTravelDistance() { return getStartPosition().getDistance(getEndPosition()); }
 	
 	public int getIndex() {
 		return index;
